+++ external/fv3fit/fv3fit/emulation/data/extra_transforms.py
@@ -9,10 +9,15 @@ import pickle
 import numpy as np
 import tensorflow as tf
 import xarray as xr
+from toolz.functoolz import curry
 from typing import Mapping, Tuple, Union
 
+from .stacking import ArrayStacker
 
+
+InputDataset = Mapping[str, Union[np.ndarray, xr.DataArray]]
 NumericContainer = Union[np.ndarray, xr.DataArray, tf.Tensor]
+ArrayDataset = Mapping[str, np.ndarray]
 StandardizeInfo = Mapping[str, Tuple[NumericContainer, NumericContainer]]
 AnyDataset = Mapping[str, NumericContainer]
 
@@ -57,3 +62,44 @@ class Standardize(DumpableTransform):
     def dump(self, path: str):
         with open(os.path.join(path, self.OUTPUT_FILE), "wb") as f:
             pickle.dump(self.std_info, f)
+
+
+@curry
+def standardize(std_info: StandardizeInfo, dataset: InputDataset):
+    """
+    Standardize data by removing the mean and scaling by a factor.
+    """
+    standardized = {}
+    for varname in dataset:
+        mean, std = std_info[varname]
+        standardized[varname] = (dataset[varname] - mean) / std
+    return standardized
+
+
+@curry
+def unstandardize(std_info: StandardizeInfo, dataset: InputDataset):
+    """
+    Unstandardize data by scaling by a factor and adding the mean.
+    """
+    unstandardized = {}
+    for varname in dataset:
+        mean, std = std_info[varname]
+        unstandardized[varname] = (dataset[varname] * std) + mean
+    return unstandardized
+
+
+@curry
+def stack_io(
+    X_stacker: ArrayStacker,
+    y_stacker: ArrayStacker,
+    dataset: ArrayDataset
+) -> Tuple[np.ndarray, np.ndarray]:
+    """
+    Use stackers to combine dataset variables along the feature dimension.
+    Groups dataset into stacked input and output data for training.
+    """
+
+    inputs = X_stacker.stack(dataset)
+    outputs = y_stacker.stack(dataset)
+
+    return inputs, outputs
