+++ external/fv3fit/fv3fit/emulation/data/stacking.py
@@ -1,3 +1,8 @@
+"""
+Old stacking infrastructure from emulation in notebooks.
+No tests yet.
+"""
+
 import yaml
 import numpy as np
 from typing import Sequence, TextIO, Mapping, Any
